Correctly retry from web socket client. Prepping web socket client for safe color mode
