+++ web-client/src/mud/prompt.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+
+export const Prompt: React.FC = () => {
+  // from https://heroicons.com/ (MIT)
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      className="h-6 w-6"
+      fill="none"
+      viewBox="0 0 24 24"
+      stroke="currentColor"
+    >
+      <path
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth={2}
+        d="M9 5l7 7-7 7"
+      />
+    </svg>
+  );
+};
